Deployments Refactored
